requirements.txt
beamngpy==1.14.1
drivebuild-client==0.40
tensorflow==2.0.0
Pillow==6.2.1